requirements.txt
# local package
-e .

# external requirements
click>=8.0.4
python-dotenv>=0.5.1

numpy>=1.22.2
opencv-python>=4.5.5.62
pillow>=9.0.1
tqdm>=4.62.3
pandas>=1.4.1
setuptools>=59.5.0
albumentations>=1.1.0
torch>=1.10.2
matplotlib>=3.5.1
torchvision>=0.11.3
wandb>=0.12.10
scikit-learn>=0.24.2
timm>=0.5.4
torchmetrics>=0.7.2